api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jwt==1.3.1
msal==1.32.0
packaging==24.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
qdrant-client==1.9.0
sentence-transformers==2.7.0
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
attrs==25.3.0
bitsandbytes==0.42.0
dataclasses-json==0.6.7
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
greenlet==3.2.3
grpcio==1.72.1
grpcio-tools==1.72.1
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpx-sse==0.4.0
huggingface-hub==0.32.4
hyperframe==6.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.2.1
numpy==2.0.1
orjson==3.10.18
ormsgpack==1.10.0
pillow==11.2.1
portalocker==2.10.1
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pydantic-settings==2.9.1
PyYAML==6.0.2
qdrant-client==1.9.0
regex==2024.11.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.7.0
sentencepiece==0.2.0
setuptools==80.9.0
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-inspect==0.9.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
